requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
distro==1.9.0
dotenv==0.9.9
google-api-core==2.30.0
google-auth==2.49.1
google-genai==1.68.0
googleapis-common-protos==1.73.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
linkify-it-py==2.1.0
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
platformdirs==4.9.4
proto-plus==1.27.1
protobuf==6.33.6
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
python-dotenv==1.2.2
requests==2.32.5
rich==14.3.3
sniffio==1.3.1
tenacity==9.1.4
textual==8.1.1
typing-extensions==4.15.0
typing-inspection==0.4.2
uc-micro-py==2.0.0
urllib3==2.6.3
websockets==16.0
